src/requirements.txt
asyncpg==0.29.0
databases[postgresql]==0.2.6
fastapi==0.109.0
sqlalchemy[asyncio]==2.0.25
uvicorn==0.27.0
pandas==2.1.4
email-validator==2.1.0.post1
alembic==1.13.1
sqladmin==0.16.0
alembic-postgresql-enum==1.0.2
PyJWT==2.8.0
cryptography==42.0.1

# dev
pytest==5.4.1
requests==2.23.0